% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{add_colorDF_theme}
\alias{add_colorDF_theme}
\title{Add a new theme}
\usage{
add_colorDF_theme(theme, id, description = NULL)
}
\arguments{
\item{theme}{a list containing style definitions}

\item{id}{an identifier (name) for the theme}

\item{description}{Description of th theme}
}
\value{
invisibly the new theme.
}
\description{
Add a new theme
}
\examples{
newtheme <- get_colorDF_theme("bw")
## Like "bw" theme, but significant p-values are red
newtheme$data.styles$pval$fg_sign <- "#FF0000"
add_colorDF_theme(newtheme, "new", "My new theme")
}
