% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{as.colorDF}
\alias{as.colorDF}
\title{Make a dataframe colorful}
\usage{
as.colorDF(x, ...)
}
\arguments{
\item{x}{a data frame or similar object (e.g. tibble)}

\item{...}{further arguments are passed to [colorDF()].}
}
\value{
a colorful data frame – identical object but with the `.style`
        attribute set.
}
\description{
Make a dataframe colorful
}
\seealso{
[df_style()] on how to modify style of the colorful data frame
}
