% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{colorDF_remove}
\alias{colorDF_remove}
\title{Strip the colorDF class and style}
\usage{
colorDF_remove(x)
}
\arguments{
\item{x}{colorful data frame}
}
\value{
the original data frame like object
}
\description{
Strip the colorDF class and style
}
