% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{colorDF_themes}
\alias{colorDF_themes}
\title{List all available themes for colorful data frames}
\usage{
colorDF_themes()
}
\value{
A character vector with the names of the themes
}
\description{
List all available themes for colorful data frames
}
\seealso{
\code{\link[=colorDF_themes_show]{colorDF_themes_show()}} for a demonstration of all themes.
}
