% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{format_col}
\alias{format_col}
\title{Format a vector using styles}
\usage{
format_col(
  x,
  col_name = NULL,
  style = NULL,
  df_style = NULL,
  format = TRUE,
  col_width = NULL,
  prefix = " ",
  min.width = 5L
)
}
\arguments{
\item{x}{a vector}

\item{col_name}{optional: a column name (see Details)}

\item{style}{A list with style definition}

\item{df_style}{style for the whole data frame}

\item{format}{Whether the vector should be formatted and aligned}

\item{col_width}{optional: width to which elements of the vector should be aligned}

\item{prefix}{prefix (column separator) to add to each element of x}

\item{min.width}{minimum width of a column}
}
\description{
Format a vector (data frame column) aligning, rounding the numbers and
adding color.
}
