% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{get_colorDF_theme}
\alias{get_colorDF_theme}
\title{Return a style defined in a theme}
\usage{
get_colorDF_theme(theme)
}
\arguments{
\item{theme}{name}
}
\value{
A list with the definitions of style
}
\description{
Return a style defined in a theme
}
\examples{
get_colorDF_theme("bw")
}
