% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{is.colorDF}
\alias{is.colorDF}
\title{Test whether an object has the class of colorDF}
\usage{
is.colorDF(x)
}
\arguments{
\item{x}{a data frame like object}
}
\value{
TRUE if the object is of colorDF class
}
\description{
Test whether an object has the class of colorDF
}
