% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_df_style}
\alias{remove_df_style}
\title{Remove the colorful dataframe style attribute}
\usage{
remove_df_style(x)
}
\arguments{
\item{x}{a colorful dataframe}
}
\value{
colorless data frame
}
\description{
Remove the colorful dataframe style attribute
}
