% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.colorDF}
\alias{summary.colorDF}
\title{Meaningful summary of data frames}
\usage{
\method{summary}{colorDF}(object, ...)
}
\arguments{
\item{object}{a data frame (possibly a color data frame)}

\item{...}{ignored}
}
\value{
A colorful data frame of class colorDF containing useful
        information on a dataframe-like object.
}
\description{
Meaningful summary of data frames
}
\details{
While this function is a summary method for objects of the colorDF
class, it can also be applied to any other dataframe-like object.
}
