% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{term_boxplot}
\alias{term_boxplot}
\title{Boxplot in a terminal}
\usage{
term_boxplot(formula, data = NULL, width = getOption("width"))
}
\arguments{
\item{formula}{a formula}

\item{data}{data frame or matrix}

\item{width}{width of the boxplot in characters}
}
\value{
invisibly return the color summary data frame used to draw the boxplot
}
\description{
Show a boxplot using characters in the terminal window
}
\examples{
term_boxplot(mpg ~ cyl, data=mtcars)
term_boxplot(Sepal.Length ~ Species, data=iris, width=70)
}
