% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{col_type<-}
\alias{col_type<-}
\alias{col_type}
\title{Set or retrieve a column type}
\usage{
col_type(x, cols = NULL) <- value

col_type(x, cols = NULL)
}
\arguments{
\item{x}{a colorful data frame}

\item{cols}{column names to set or retrieve}

\item{value}{character vector with column types}
}
\description{
Set or retrieve a column type of a colorful data frame
}
\examples{
mc <- colorDF(mtcars)
col_type(mc, "gear") <- "factor"
col_type(mc, "gear")
col_type(mc) <- list(gear="factor", cyl="integer")
}
