% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{colorDF}
\alias{colorDF}
\title{Make a dataframe colorful}
\usage{
colorDF(x, theme = NULL)
}
\arguments{
\item{x}{a data frame or similar object (e.g. tibble)}

\item{theme}{Which theme to use}
}
\value{
a colorful data frame – identical object but with the `.style`
        attribute set.
}
\description{
Make a dataframe colorful
}
\seealso{
[colorDF_themes()] to list all themes; [colorDF_themes_show()]
         to view all themes.

[df_style()] on how to modify style of the colorful data frame
}
