% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{get_colorDF_theme}
\alias{get_colorDF_theme}
\title{Return a style defined in a theme}
\usage{
get_colorDF_theme(theme)
}
\arguments{
\item{theme}{name}
}
\value{
A list with the definitions of style
}
\description{
Return a style defined in a theme
}
\examples{
get_colorDF_theme("bw")

## use it to change the style of a colorDF
foo <- colorDF(mtcars)
df_style(foo) <- get_colorDF_theme("wb")
## Slightly modify the style
df_style(foo)$type.styles$pval$fg_sign <- "red"
}
