% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.colorDF}
\alias{print.colorDF}
\title{Print method for colorful data frames}
\usage{
\method{print}{colorDF}(
  x,
  n = 20,
  width = getOption("width"),
  row.names = TRUE,
  highlight = NULL,
  ...
)
}
\arguments{
\item{x}{a colorful data frame (object with class colorDF)}

\item{n}{Number of rows to show (default=20, use Inf to show all)}

\item{width}{number of characters that the data frame should span on output}

\item{row.names}{if TRUE (default), row names will be shown on output}

\item{highlight}{a logical vector indicating which rows to highlight}

\item{...}{further arguments are ignored}
}
\description{
Print method for colorful data frames
}
\seealso{
[df_style()] on how to modify colorful data frames
}
