% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_df_style}
\alias{remove_df_style}
\title{Remove the colorful dataframe style attribute}
\usage{
remove_df_style(x)
}
\arguments{
\item{x}{a colorful dataframe}
}
\value{
colorless data frame
}
\description{
Remove the colorful dataframe style attribute
}
\details{
Strips the color data frame style, but leaves the .coltp and class
intact.
}
\seealso{
To completely remove the colorDF class and attributes, use \code{\link[=uncolor]{uncolor()}}
}
